fix hell memory usage when app has many users
bug cause:
server/src/libs/room.rs:279
Every time room receives a PushSegment message from user,
it sends the message to other users.
To accomplish this, send_all_except_sender function is invoked, which needs to clone data to send.
Average size of a line segment is ~500 bytes, it is pretty enough to fill the memory if we have many concurrent senders.

also:
- add stress-test
- removed realtime drawing to improve perfomance